# contracts/openapi/horizon-api.yaml
openapi: 3.1.0
info:
  title: Horizon Public API
  version: 1.0.0
  description: |
    Публичный API платформы Horizon — обнаружение зарождающихся научно-технологических
    трендов (слабых сигналов) в открытых источниках.

    Контракт **first-class**: спецификация является источником истины, реализация проверяется
    контрактными тестами (`spring-boot-starter-web` + `swagger-request-validator`, Schemathesis для Python).

    Все ошибки — RFC 9457 `application/problem+json` (ADR-0013).
    Входа нет: организация одна, и API открыт так же, как интерфейс. Служебные вызовы между
    сервисами защищены общим секретом (`X-Internal-Token`) и в этот контракт не входят.
  contact:
    name: Horizon Team
    url: https://horizon.dev
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Локальная среда (gateway)
  - url: https://horizon.example.bank
    description: Промышленная среда
...
paths:
...
  /api/v1/reports/{reportId}/trends/{trendKey}/feedback:
    put:
      tags: [Feedback]
      summary: Оценить тренд
      operationId: submitTrendFeedback
      parameters:
        - $ref: '#/components/parameters/ReportId'
        - name: trendKey
          in: path
          required: true
          schema: { type: string, maxLength: 160 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TrendFeedbackRequest' }
      responses:
        '200':
          description: Оценка сохранена
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TrendFeedbackView' }
        '400': { $ref: '#/components/responses/ValidationProblem' }
        '404': { $ref: '#/components/responses/NotFoundProblem' }
    delete:
      tags: [Feedback]
      summary: Снять свою пометку с темы
      description: >-
        Отдельная операция, а не вердикт «полезно». Пометка «не технология» стала действием с
        последствиями: движок убирает такую тему до отбора в ТОП-N, и в отчёте её больше нет.
        Отменить это, объявив тему полезной, значит сделать утверждение, которого аналитик не делал,
        — а его унаследует следующий отчёт. Снимается только собственная пометка: ключи, которых
        аналитик не помечал, отвергаются, иначе появился бы способ писать в разметку произвольные
        строки.
      operationId: withdrawTrendFeedback
      parameters:
        - $ref: '#/components/parameters/ReportId'
        - name: trendKey
          in: path
          required: true
          schema: { type: string, maxLength: 160 }
      responses:
        '204': { description: Пометка снята }
        '404': { $ref: '#/components/responses/NotFoundProblem' }

  # ──────────────────────────── SavedDomains ──────────────────────────
...
  /api/v1/sources/{sourceId}:
    patch:
      tags: [Sources]
      summary: Изменить настройки источника
      description: Требуется роль `RESEARCHER` или `ADMIN`.
      operationId: updateSource
      parameters:
        - { name: sourceId, in: path, required: true, schema: { type: string, maxLength: 48 } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateSourceRequest' }
      responses:
        '200':
          description: Обновлено
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SourceView' }
        '404': { $ref: '#/components/responses/NotFoundProblem' }
  /api/v1/sources/{sourceId}/runs:
    post:
      tags: [Sources]
      summary: Запустить прогон сбора вручную
      operationId: triggerIngestionRun
      parameters:
        - { name: sourceId, in: path, required: true, schema: { type: string, maxLength: 48 } }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TriggerIngestionRequest' }
      responses:
        '202':
          description: Прогон запущен
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IngestionRunView' }
        '409':
          description: Прогон по этому источнику уже выполняется
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /api/v1/ingestion-runs:
    get:
      tags: [Sources]
      summary: История прогонов сбора
      operationId: listIngestionRuns
      parameters:
        - { name: sourceId, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, minimum: 0, default: 0 } }
        - { name: size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200':
          description: Страница прогонов
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IngestionRunPage' }
components:
  parameters:
...
    ReportId:
      name: reportId
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    ValidationProblem:
      description: Ошибка валидации
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFoundProblem:
      description: Не найдено
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 Problem Details
      required: [type, title, status]
      properties:
        type: { type: string, format: uri, examples: ['https://horizon.dev/problems/validation-error'] }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        traceId: { type: string }
        timestamp: { type: string, format: date-time }
        retryable: { type: boolean }
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
              rejectedValue: {}
    PageMeta:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
    SourceClass:
      type: string
      enum: [PREPRINT, JOURNAL_ARTICLE, PATENT, CODE_REPOSITORY, NEWS, ANALYST_REPORT, STANDARD]
...
    TrendFeedbackRequest:
      type: object
      required: [verdict]
      properties:
        verdict: { type: string, enum: [RELEVANT, NOISE, ALREADY_KNOWN] }
        comment: { type: string, maxLength: 2000 }
    TrendFeedbackView:
      type: object
      required: [verdict, createdAt]
      properties:
        verdict: { type: string, enum: [RELEVANT, NOISE, ALREADY_KNOWN] }
        comment: { type: string }
        createdAt: { type: string, format: date-time }
        carried:
          type: boolean
          default: false
          description: >-
            Оценка дана не в этом отчёте, а в другом по тому же направлению — как правило, в более
            ранней версии, ради которой перенос и сделан: отчёты неизменяемы, пересчёт создаёт новый,
            и без переноса разметка аналитика пропадала бы каждый квартал. Ограничения «строго
            раньше» намеренно нет: сказано ровно то, что проверено, — это другой отчёт того же
            направления.


            Помечена как перенесённая, а не выдана за здешнюю: поставленная продуктом от имени
            аналитика пометка неотличима от его собственной, и он нашёл бы своё имя под суждением,
            которого не высказывал. Подтверждается обычным `PUT`.
...
    SourceView:
      type: object
      required: [id, displayName, sourceClass, enabled]
      properties:
        id: { type: string }
        displayName: { type: string }
        sourceClass: { $ref: '#/components/schemas/SourceClass' }
        enabled: { type: boolean }
        requiresApiKey: { type: boolean }
        apiKeyConfigured: { type: boolean }
        rateLimitPerMinute: { type: integer }
        documentCount: { type: integer, format: int64 }
        lastRun: { $ref: '#/components/schemas/IngestionRunView' }
...
    UpdateSourceRequest:
      type: object
      properties:
        enabled: { type: boolean }
        rateLimitPerMinute: { type: integer, minimum: 1, maximum: 6000 }
    TriggerIngestionRequest:
      type: object
      properties:
        mode: { type: string, enum: [INCREMENTAL, BACKFILL], default: INCREMENTAL }
        query: { type: string, maxLength: 200 }
        windowFrom: { type: string, format: date }
        windowTo: { type: string, format: date }
    IngestionRunView:
      type: object
      required: [id, sourceId, status, startedAt]
      properties:
        id: { type: string, format: uuid }
        sourceId: { type: string }
        mode: { type: string, enum: [INCREMENTAL, BACKFILL, ON_DEMAND] }
        status: { type: string, enum: [RUNNING, COMPLETED, PARTIAL, FAILED] }
        documentsFetched: { type: integer }
        documentsCreated: { type: integer }
        documentsDuplicate: { type: integer }
        documentsRejected: { type: integer }
        errorCode: { type: [string, 'null'] }
        errorMessage: { type: [string, 'null'] }
        startedAt: { type: string, format: date-time }
        finishedAt: { type: [string, 'null'], format: date-time }
    IngestionRunPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [content]
          properties:
            content:
              type: array
              items: { $ref: '#/components/schemas/IngestionRunView' }